perf: admin dashobard pages
