+++ app/Http/Controllers/Admin/DashboardController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+class DashboardController extends Controller
+{
+    public function __invoke()
+    {
+        return view('pages.admin.dashboard', [
+            'title' => 'Beranda',
+            'stats' => $this->orderStats(),
+            'stockCards' => $this->stockCards(),
+            'filterCategory' => null,
+        ]);
+    }
+}
