+++ resources/views/pages/admin/_partials/orders-manager.blade.php
@@ -1,0 +1,355 @@
+@php $isTicket = $filterCategory === 'Tiket'; @endphp
+@if ($filterCategory)
+<div class="grid grid-cols-2 gap-3 sm:grid-cols-3 lg:grid-cols-4" data-stats-root>
+    <div class="rounded-xl border border-mercury bg-white p-4">
+        <p class="text-[10px] font-semibold uppercase tracking-wider text-onyx">Total Pesanan</p>
+        <p class="mt-1.5 text-xl font-black text-foreground" data-stat="total">{{ number_format($stats['total']) }}</p>
+    </div>
+    <div class="rounded-xl border border-mercury bg-white p-4">
+        <p class="text-[10px] font-semibold uppercase tracking-wider text-onyx">Menunggu Pembayaran</p>
+        <p class="mt-1.5 text-xl font-black text-amber-600" data-stat="pending">{{ number_format($stats['pending']) }}</p>
+    </div>
+    <div class="group relative rounded-xl border border-mercury bg-white p-4">
+        <div class="flex items-start justify-between gap-2">
+            <p class="text-[10px] font-semibold uppercase tracking-wider text-onyx">Pembayaran Diterima</p>
+            @unless ($isTicket)
+            <button type="button" class="-mr-1 -mt-1 shrink-0 rounded-md p-1 text-onyx/50 transition hover:text-primary focus:text-primary focus:outline-none" aria-label="Rincian DP &amp; Lunas">
+                <i class="ri-information-line text-sm"></i>
+            </button>
+            @endunless
+        </div>
+        <p class="mt-1.5 text-xl font-black text-emerald-600" data-stat="confirmed">{{ number_format($stats['confirmed']) }}</p>
+
+        @unless ($isTicket)
+        <div class="invisible absolute right-3 top-12 z-20 w-48 origin-top-right -translate-y-1 scale-95 rounded-xl border border-mercury bg-white p-3 opacity-0 shadow-xl ring-1 ring-black/5 transition duration-150 group-hover:visible group-hover:translate-y-0 group-hover:scale-100 group-hover:opacity-100 group-focus-within:visible group-focus-within:translate-y-0 group-focus-within:scale-100 group-focus-within:opacity-100">
+            <p class="text-[10px] font-bold uppercase tracking-wider text-onyx">Rincian Pembayaran</p>
+            <div class="mt-2 space-y-2">
+                <div class="flex items-center justify-between gap-2">
+                    <span class="inline-flex items-center gap-1.5 text-[11px] font-semibold text-amber-700"><i class="ri-hourglass-line"></i> DP (belum lunas)</span>
+                    <span class="text-[13px] font-black text-amber-700" data-stat="verified">{{ number_format($stats['verified']) }}</span>
+                </div>
+                <div class="flex items-center justify-between gap-2">
+                    <span class="inline-flex items-center gap-1.5 text-[11px] font-semibold text-teal-700"><i class="ri-checkbox-circle-line"></i> Lunas</span>
+                    <span class="text-[13px] font-black text-teal-700" data-stat="settled">{{ number_format($stats['settled']) }}</span>
+                </div>
+            </div>
+        </div>
+        @endunless
+    </div>
+    <div class="col-span-2 rounded-xl bg-linear-to-br from-primary via-primary-light to-primary-lighter p-4 text-white sm:col-span-3 lg:col-span-1">
+        <p class="text-[10px] font-semibold uppercase tracking-wider text-white/80">Total Pendapatan</p>
+        <p class="mt-1.5 text-xl font-black" data-stat="revenue">Rp{{ number_format($stats['revenue'], 0, ',', '.') }}</p>
+    </div>
+</div>
+
+@if (!empty($stockCards))
+<div class="mt-3 flex flex-col gap-3 empty:mt-0" data-stock-root>
+    @include('pages.admin._partials.stock-cards', ['stockCards' => $stockCards])
+</div>
+@endif
+
+<div class="mt-6">
+    <div class="mb-3">
+        <h3 class="text-base font-bold text-foreground">{{ $filterCategory === 'Tiket' ? 'Daftar Pesanan Tiket' : ($filterCategory === 'Merchandise' ? 'Daftar Pesanan Merchandise' : 'Daftar Semua Pesanan') }}</h3>
+        <p class="text-xs text-onyx">Kelola seluruh pesanan dari pelanggan.</p>
+    </div>
+
+    <div class="overflow-hidden rounded-xl border border-mercury bg-white shadow-sm"
+        data-orders-root
+        data-filter-category="{{ $filterCategory ?? '' }}"
+        data-data-url="{{ route('admin.orders.data') }}"
+        data-export-url="{{ route('admin.orders.export') }}"
+        data-detail-url="{{ url('admin/orders/__ORDER__') }}"
+        data-status-url="{{ url('admin/orders/__ORDER__/status') }}"
+        data-sync-payment-url="{{ url('admin/orders/__ORDER__/sync-payment') }}"
+        data-settlement-verify-url="{{ url('admin/orders/__ORDER__/settlement-verify') }}"
+        data-shipping-url="{{ url('admin/orders/__ORDER__/shipping') }}"
+        data-pickup-url="{{ url('admin/orders/__ORDER__/pickup') }}"
+        data-payment-proof-url="{{ url('admin/orders/__ORDER__/payment-proof') }}"
+        data-delete-url="{{ url('admin/orders/__ORDER__') }}">
+        <div class="orders-filters grid grid-cols-1 gap-3 border-b border-mercury bg-skull/40 p-4 {{ $isTicket ? 'sm:grid-cols-2' : 'sm:grid-cols-2 lg:grid-cols-4' }}">
+            <div>
+                <label class="mb-1.5 block text-[10px] font-bold uppercase tracking-wider text-onyx">Tanggal Pemesanan</label>
+                <div class="jpw-flatpickr-wrap">
+                    <input type="text" class="jpw-flatpickr" data-filter="date_range" placeholder="Pilih rentang tanggal…" autocomplete="off" />
+                    <button type="button" class="jpw-flatpickr-clear" data-filter-date-clear aria-label="Hapus tanggal" hidden>
+                        <i class="ri-close-line"></i>
+                    </button>
+                </div>
+            </div>
+            @unless ($isTicket)
+            <div>
+                <label class="mb-1.5 block text-[10px] font-bold uppercase tracking-wider text-onyx">Metode Pengiriman</label>
+                <select class="orders-filter w-full" data-filter="shipping_method">
+                    <option value="">Semua Pengiriman</option>
+                    <option value="kirim">Dikirim (Kurir)</option>
+                    <option value="pickup">Ambil di Tempat</option>
+                </select>
+            </div>
+            <div>
+                <label class="mb-1.5 block text-[10px] font-bold uppercase tracking-wider text-onyx">Jenis Pembayaran</label>
+                <select class="orders-filter w-full" data-filter="payment_type">
+                    <option value="">Semua Tipe</option>
+                    <option value="dp">DP (50%)</option>
+                    <option value="full">Bayar Lunas</option>
+                </select>
+            </div>
+            @endunless
+            <div>
+                <label class="mb-1.5 block text-[10px] font-bold uppercase tracking-wider text-onyx">Status Pesanan</label>
+                <select class="orders-filter w-full" data-filter="status">
+                    <option value="">Semua Status</option>
+                    <option value="pending">Menunggu Pembayaran</option>
+                    <option value="verified">Pembayaran Diterima</option>
+                    @unless ($isTicket)
+                    <option value="paid">Pembayaran Lunas</option>
+                    <option value="shipped">Pesanan Dikirim / Siap Diambil</option>
+                    @endunless
+                    <option value="completed">Pesanan Selesai</option>
+                    <option value="cancelled">Pesanan Dibatalkan</option>
+                </select>
+            </div>
+            <div class="flex items-end gap-2 {{ $isTicket ? 'sm:col-span-2' : 'sm:col-span-2 lg:col-span-4' }}">
+                <button type="button" data-filter-reset
+                    class="inline-flex h-10 flex-1 items-center justify-center gap-1.5 rounded-lg border border-mercury bg-white px-4 text-[13px] font-semibold text-foreground transition hover:bg-skull">
+                    <i class="ri-refresh-line"></i>
+                    Reset Filter
+                </button>
+                <button type="button" data-export-orders
+                    class="inline-flex h-10 flex-1 items-center justify-center gap-1.5 rounded-lg border border-emerald-200 bg-emerald-50 px-4 text-[13px] font-semibold text-emerald-700 transition hover:bg-emerald-100">
+                    <i class="ri-file-excel-2-line"></i>
+                    Ekspor
+                </button>
+            </div>
+        </div>
+        <div class="overflow-x-auto">
+            <table id="orders-table" class="display w-full">
+                <thead>
+                    <tr>
+                        <th class="whitespace-nowrap">ID Pesanan</th>
+                        <th class="whitespace-nowrap">Tanggal</th>
+                        <th class="whitespace-nowrap">Pelanggan</th>
+                        <th class="whitespace-nowrap text-center">Item</th>
+                        <th class="whitespace-nowrap text-right">Total</th>
+                        <th class="whitespace-nowrap">Pembayaran</th>
+                        <th class="whitespace-nowrap">Status</th>
+                        <th class="whitespace-nowrap text-right">Aksi</th>
+                    </tr>
+                </thead>
+                <tbody></tbody>
+            </table>
+        </div>
+    </div>
+</div>
+
+<div id="order-detail-modal" class="order-modal" hidden aria-hidden="true">
+    <div class="order-modal__backdrop" data-modal-close></div>
+    <div class="order-modal__dialog" role="dialog" aria-modal="true" aria-labelledby="order-detail-modal-title">
+        <button type="button" class="order-modal__close" data-modal-close aria-label="Tutup">
+            <i class="ri-close-line"></i>
+        </button>
+        <h2 id="order-detail-modal-title" class="sr-only">Detail Pesanan</h2>
+        <div class="order-modal__body" data-modal-content></div>
+    </div>
+</div>
+
+<div id="sync-payment-modal" class="order-modal" hidden aria-hidden="true">
+    <div class="order-modal__backdrop" data-sync-close></div>
+    <div class="order-modal__dialog" role="dialog" aria-modal="true" aria-labelledby="sync-payment-modal-title">
+        <button type="button" class="order-modal__close" data-sync-close aria-label="Tutup">
+            <i class="ri-close-line"></i>
+        </button>
+        <div class="order-modal__body">
+            <div class="detail-hero">
+                <div class="detail-hero__bg"></div>
+                <div class="relative flex flex-col gap-2 pr-12">
+                    <span class="detail-chip detail-chip--glass w-fit font-mono"><i class="ri-refresh-line"></i> <span data-sync-order-id>—</span></span>
+                    <h2 id="sync-payment-modal-title" class="text-base font-black leading-tight text-white">Sinkronisasi Pembayaran</h2>
+                    <p class="text-[12px] leading-relaxed text-white/85">Sesuaikan total pembayaran sesuai nominal yang sudah diterima dari pembeli.</p>
+                </div>
+            </div>
+
+            <form data-sync-form class="flex flex-col gap-4 px-5 py-5">
+                <div class="rounded-2xl border border-mercury bg-skull/40 p-4">
+                    <div class="grid grid-cols-3 divide-x divide-mercury">
+                        <div class="flex flex-col items-start pr-3">
+                            <span class="text-[10px] font-bold uppercase tracking-wider text-onyx">Subtotal</span>
+                            <span class="mt-1 text-[14px] font-black leading-tight text-foreground" data-sync-subtotal>Rp0</span>
+                        </div>
+                        <div class="flex flex-col items-start px-3">
+                            <span class="text-[10px] font-bold uppercase tracking-wider text-onyx">Saat Ini</span>
+                            <span class="mt-1 text-[14px] font-black leading-tight text-foreground" data-sync-current>Rp0</span>
+                        </div>
+                        <div class="flex flex-col items-start pl-3">
+                            <span class="text-[10px] font-bold uppercase tracking-wider text-amber-700">Sisa Setelahnya</span>
+                            <span class="mt-1 text-[14px] font-black leading-tight text-amber-800" data-sync-remaining>Rp0</span>
+                        </div>
+                    </div>
+
+                    <div class="mt-3">
+                        <div class="relative h-2 w-full overflow-hidden rounded-full bg-white ring-1 ring-mercury">
+                            <div class="absolute inset-y-0 left-0 bg-linear-to-r from-primary to-primary-light transition-[width] duration-200" style="width: 0%" data-sync-progress></div>
+                        </div>
+                        <div class="mt-1.5 flex items-center justify-between text-[10px] font-bold text-onyx">
+                            <span><span data-sync-percent>0</span>% dari subtotal</span>
+                            <span data-sync-percent-label class="text-onyx">Belum dibayar</span>
+                        </div>
+                    </div>
+                </div>
+
+                <div>
+                    <div class="mb-1.5 flex items-end justify-between">
+                        <label for="sync-amount-input" class="text-[11px] font-bold uppercase tracking-wider text-onyx">Total Pembayaran Baru</label>
+                        <span class="text-[10px] font-semibold text-onyx">Maks. <span data-sync-max>Rp0</span></span>
+                    </div>
+                    <div class="group relative">
+                        <span class="pointer-events-none absolute inset-y-0 left-0 flex w-12 items-center justify-center border-r border-mercury text-[13px] font-bold text-onyx">Rp</span>
+                        <input id="sync-amount-input" type="text" inputmode="numeric" autocomplete="off" data-sync-input
+                            class="h-12 w-full rounded-xl border border-mercury bg-white pl-14 pr-4 text-[16px] font-black tabular-nums text-foreground outline-none transition focus:border-primary focus:ring-2 focus:ring-primary/20" />
+                    </div>
+                    <p class="mt-1.5 hidden text-[11px] font-semibold text-red-600" data-sync-error></p>
+                </div>
+
+                <div>
+                    <div class="mb-1.5 text-[11px] font-bold uppercase tracking-wider text-onyx">Isi Cepat</div>
+                    <div class="grid grid-cols-3 gap-2">
+                        <button type="button" data-sync-preset="50" class="inline-flex h-9 items-center justify-center gap-1 rounded-lg border border-mercury bg-white px-2 text-[11px] font-bold text-foreground transition hover:border-primary hover:bg-primary-softer hover:text-primary"><i class="ri-percent-line text-[12px]"></i> 50% DP</button>
+                        <button type="button" data-sync-preset="75" class="inline-flex h-9 items-center justify-center gap-1 rounded-lg border border-mercury bg-white px-2 text-[11px] font-bold text-foreground transition hover:border-primary hover:bg-primary-softer hover:text-primary"><i class="ri-percent-line text-[12px]"></i> 75%</button>
+                        <button type="button" data-sync-preset="100" class="inline-flex h-9 items-center justify-center gap-1 rounded-lg border border-emerald-200 bg-emerald-50 px-2 text-[11px] font-bold text-emerald-700 transition hover:bg-emerald-100"><i class="ri-checkbox-circle-line text-[12px]"></i> Lunas</button>
+                    </div>
+                </div>
+
+                <div class="-mx-5 -mb-5 mt-1 flex items-center justify-end gap-2 border-t border-mercury bg-skull/40 px-5 py-3">
+                    <button type="button" data-sync-close class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg border border-mercury bg-white px-4 text-[13px] font-semibold text-foreground transition hover:bg-skull">Batal</button>
+                    <button type="submit" data-sync-submit class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg bg-primary px-4 text-[13px] font-bold text-white shadow-sm transition hover:bg-primary-light disabled:cursor-not-allowed disabled:opacity-60">
+                        <i class="ri-save-3-line"></i> Simpan Perubahan
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+<div id="shipping-modal" class="order-modal" hidden aria-hidden="true">
+    <div class="order-modal__backdrop" data-shipping-close></div>
+    <div class="order-modal__dialog" role="dialog" aria-modal="true" aria-labelledby="shipping-modal-title">
+        <button type="button" class="order-modal__close" data-shipping-close aria-label="Tutup">
+            <i class="ri-close-line"></i>
+        </button>
+        <div class="order-modal__body">
+            <div class="detail-hero">
+                <div class="detail-hero__bg"></div>
+                <div class="relative flex flex-col gap-2 pr-12">
+                    <span class="detail-chip detail-chip--glass w-fit font-mono"><i class="ri-truck-line"></i> <span data-shipping-order-id>—</span></span>
+                    <h2 id="shipping-modal-title" data-shipping-title class="text-base font-black leading-tight text-white">Nomor Resi Pengiriman</h2>
+                    <p data-shipping-subtitle class="text-[12px] leading-relaxed text-white/85">Masukkan nomor resi JNT Express. Resi wajib diisi sebelum pesanan kirim bisa ditandai selesai.</p>
+                </div>
+            </div>
+
+            <form data-shipping-form class="flex flex-col gap-4 px-5 py-5">
+                <div>
+                    <label for="shipping-tracking-input" class="mb-1.5 block text-[11px] font-bold uppercase tracking-wider text-onyx">Nomor Resi</label>
+                    <div class="group relative">
+                        <span class="pointer-events-none absolute inset-y-0 left-0 flex w-12 items-center justify-center border-r border-mercury text-[15px] text-onyx"><i class="ri-barcode-line"></i></span>
+                        <input id="shipping-tracking-input" type="text" autocomplete="off" maxlength="100" data-shipping-input placeholder="Cth. JX1234567890"
+                            class="h-12 w-full rounded-xl border border-mercury bg-white pl-14 pr-4 text-[15px] font-bold tracking-wide text-foreground outline-none transition focus:border-primary focus:ring-2 focus:ring-primary/20" />
+                    </div>
+                    <p class="mt-1.5 hidden text-[11px] font-semibold text-red-600" data-shipping-error></p>
+                </div>
+
+                <div class="-mx-5 -mb-5 mt-1 flex items-center justify-end gap-2 border-t border-mercury bg-skull/40 px-5 py-3">
+                    <button type="button" data-shipping-close class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg border border-mercury bg-white px-4 text-[13px] font-semibold text-foreground transition hover:bg-skull">Batal</button>
+                    <button type="submit" data-shipping-submit class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg bg-primary px-4 text-[13px] font-bold text-white shadow-sm transition hover:bg-primary-light disabled:cursor-not-allowed disabled:opacity-60">
+                        <i class="ri-save-3-line"></i> Simpan Resi
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+<div id="pickup-modal" class="order-modal" hidden aria-hidden="true">
+    <div class="order-modal__backdrop" data-pickup-close></div>
+    <div class="order-modal__dialog" role="dialog" aria-modal="true" aria-labelledby="pickup-modal-title">
+        <button type="button" class="order-modal__close" data-pickup-close aria-label="Tutup">
+            <i class="ri-close-line"></i>
+        </button>
+        <div class="order-modal__body">
+            <div class="detail-hero">
+                <div class="detail-hero__bg"></div>
+                <div class="relative flex flex-col gap-2 pr-12">
+                    <span class="detail-chip detail-chip--glass w-fit font-mono"><i class="ri-store-2-line"></i> <span data-pickup-order-id>—</span></span>
+                    <h2 id="pickup-modal-title" data-pickup-title class="text-base font-black leading-tight text-white">Titik Pengambilan</h2>
+                    <p data-pickup-subtitle class="text-[12px] leading-relaxed text-white/85">Tentukan alamat & kontak pengurus untuk pesanan ini. Info ini dikirim ke pembeli saat ditandai siap diambil.</p>
+                </div>
+            </div>
+
+            <form data-pickup-form class="flex flex-col gap-4 px-5 py-5">
+                <div>
+                    <label for="pickup-address-input" class="mb-1.5 block text-[11px] font-bold uppercase tracking-wider text-onyx">Alamat / Titik Temu</label>
+                    <textarea id="pickup-address-input" rows="2" maxlength="255" autocomplete="off" data-pickup-address placeholder="Cth. Jl. Soedirman No.10, depan Alun-alun Purwokerto"
+                        class="w-full rounded-xl border border-mercury bg-white px-4 py-2.5 text-[14px] text-foreground outline-none transition focus:border-primary focus:ring-2 focus:ring-primary/20"></textarea>
+                </div>
+                <div>
+                    <label for="pickup-contact-name-input" class="mb-1.5 block text-[11px] font-bold uppercase tracking-wider text-onyx">Nama Pengurus</label>
+                    <input id="pickup-contact-name-input" type="text" maxlength="100" autocomplete="off" data-pickup-contact-name placeholder="Cth. Budi"
+                        class="h-12 w-full rounded-xl border border-mercury bg-white px-4 text-[14px] text-foreground outline-none transition focus:border-primary focus:ring-2 focus:ring-primary/20" />
+                </div>
+                <div>
+                    <label for="pickup-contact-phone-input" class="mb-1.5 block text-[11px] font-bold uppercase tracking-wider text-onyx">Nomor Kontak (WhatsApp)</label>
+                    <input id="pickup-contact-phone-input" type="text" inputmode="numeric" maxlength="30" autocomplete="off" data-pickup-contact-phone placeholder="Cth. 6281234567890"
+                        class="h-12 w-full rounded-xl border border-mercury bg-white px-4 text-[14px] tabular-nums text-foreground outline-none transition focus:border-primary focus:ring-2 focus:ring-primary/20" />
+                    <p class="mt-1 text-[11px] text-onyx">Format internasional tanpa "+" (mis. 6281234567890).</p>
+                </div>
+                <p class="hidden text-[11px] font-semibold text-red-600" data-pickup-error></p>
+
+                <div class="-mx-5 -mb-5 mt-1 flex items-center justify-end gap-2 border-t border-mercury bg-skull/40 px-5 py-3">
+                    <button type="button" data-pickup-close class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg border border-mercury bg-white px-4 text-[13px] font-semibold text-foreground transition hover:bg-skull">Batal</button>
+                    <button type="submit" data-pickup-submit class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg bg-primary px-4 text-[13px] font-bold text-white shadow-sm transition hover:bg-primary-light disabled:cursor-not-allowed disabled:opacity-60">
+                        <i class="ri-save-3-line"></i> Simpan
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+
+<div id="payment-proof-modal" class="order-modal" hidden aria-hidden="true">
+    <div class="order-modal__backdrop" data-proof-close></div>
+    <div class="order-modal__dialog" role="dialog" aria-modal="true" aria-labelledby="payment-proof-modal-title">
+        <button type="button" class="order-modal__close" data-proof-close aria-label="Tutup">
+            <i class="ri-close-line"></i>
+        </button>
+        <div class="order-modal__body">
+            <div class="detail-hero">
+                <div class="detail-hero__bg"></div>
+                <div class="relative flex flex-col gap-2 pr-12">
+                    <span class="detail-chip detail-chip--glass w-fit font-mono"><i class="ri-receipt-line"></i> <span data-proof-order-id>—</span></span>
+                    <h2 id="payment-proof-modal-title" data-proof-title class="text-base font-black leading-tight text-white">Bukti Transfer</h2>
+                    <p data-proof-subtitle class="text-[12px] leading-relaxed text-white/85">Unggah bukti transfer pembayaran untuk pesanan ini.</p>
+                </div>
+            </div>
+
+            <form data-proof-form class="flex flex-col gap-4 px-5 py-5">
+                <div>
+                    <label class="mb-1.5 block text-[11px] font-bold uppercase tracking-wider text-onyx">File Bukti Transfer</label>
+                    <label class="flex cursor-pointer flex-col items-center justify-center gap-2 rounded-xl border-2 border-dashed border-mercury bg-skull/40 px-4 py-6 text-center transition hover:border-primary hover:bg-primary-softer">
+                        <i class="ri-upload-cloud-2-line text-2xl text-onyx"></i>
+                        <span data-proof-filename class="text-[12px] font-semibold text-onyx">Klik untuk pilih file</span>
+                        <span class="text-[10px] text-onyx/70">JPG · PNG · WEBP · PDF — maks. 5 MB</span>
+                        <input type="file" accept=".jpg,.jpeg,.png,.webp,.pdf,image/*,application/pdf" data-proof-input class="hidden" />
+                    </label>
+                    <p class="mt-1.5 hidden text-[11px] font-semibold text-red-600" data-proof-error></p>
+                </div>
+
+                <div class="-mx-5 -mb-5 mt-1 flex items-center justify-end gap-2 border-t border-mercury bg-skull/40 px-5 py-3">
+                    <button type="button" data-proof-close class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg border border-mercury bg-white px-4 text-[13px] font-semibold text-foreground transition hover:bg-skull">Batal</button>
+                    <button type="submit" data-proof-submit class="inline-flex h-10 items-center justify-center gap-1.5 rounded-lg bg-primary px-4 text-[13px] font-bold text-white shadow-sm transition hover:bg-primary-light disabled:cursor-not-allowed disabled:opacity-60">
+                        <i class="ri-upload-2-line"></i> Unggah
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+@endif
