+++ resources/views/pages/admin/merchandise/index.blade.php
@@ -1,6 +1,6 @@
 @extends('layouts.admin')
 
-@section('heading', $title ?? 'Dashboard')
+@section('heading', $title ?? 'Pesanan Merchandise')
 
 @push('styles')
     @vite([
@@ -12,10 +12,6 @@
 @endpush
 
 @section('content')
-    <div class="mb-6">
-        <h2 class="text-xl font-semibold text-gray-900">Selamat datang, {{ auth()->user()->name }}.</h2>
-        <p class="text-sm text-gray-600">Pilih menu <span class="font-semibold">Pesanan Tiket</span> atau <span class="font-semibold">Pesanan Merchandise</span> di sisi kiri untuk mengelola pesanan.</p>
-    </div>
     @include('pages.admin._partials.orders-manager')
 @endsection
